chore: type corrections and more mocks
