+++ src/constants/endpoints.ts
@@ -2,5 +2,6 @@ const GET_BILLS = "/bills";
 const SEARCH_BILLS = "/bill/search";
 const GET_CONTACTS = "/contacts/list";
 const SEARCH_CONTACTS = "/contacts/search";
+const EDIT_CONTACT = "/contacts/edit";
 
-export { GET_BILLS, SEARCH_BILLS, GET_CONTACTS, SEARCH_CONTACTS };
+export { GET_BILLS, SEARCH_BILLS, GET_CONTACTS, SEARCH_CONTACTS, EDIT_CONTACT };
