+++ src/i18n/source.json
@@ -5,7 +5,6 @@
   "Against the sum of": "Against the sum of",
   "Against this bill of exchange": "Against this bill of exchange",
   "All": "All",
-  "Are you sure": "Are you sure?",
   "Authorised signer": "Authorised signer",
   "Balances": "Balances",
   "Bill issuer": "Bill issuer",
@@ -45,7 +44,6 @@
   "Draw bill": "Draw bill",
   "E-bills": "E-bills",
   "Edit": "Edit",
-  "Edit contact": "Edit contact",
   "Edit issue": "Edit issue",
   "Email address": "Email address",
   "Endorse to the order of": "Endorse to the order of",
@@ -108,7 +106,6 @@
   "Required information": "Required information",
   "Restore identity": "Restore identity",
   "Restore with seed phrase": "Restore with seed phrase",
-  "Save": "Save",
   "Search": "Search",
   "Search field for home page": "Search...",
   "Search for bills": "Search for bills...",
@@ -136,7 +133,6 @@
   "Upload document": "Upload document",
   "Upload proof of registration": "Upload proof of registration",
   "We have successfully created your identity and you can now use the app": "We have successfully created your identity and you can now use the app",
-  "You can always re-add a contact later": "You can always re-add a contact later",
   "You can now view the bill in your overview of bills": "You can now view the bill in your overview of bills",
   "authorizedSigners.title": "Authorized Signers",
   "bill.actions.actionsList": "Endorse",
@@ -183,9 +179,11 @@
   "bill.mint.requestToMint": "Request to mint",
   "bill.mint.selectQuote": "Select a quote",
   "bill.mint.title": "Mint quotes",
-  "contact.action.delete.cancel": "Cancel",
-  "contact.action.delete.confirm": "Delete",
-  "contact.delete": "Delete contact",
+  "contact.view.delete": "Delete contact",
+  "contact.view.delete.cancel": "Cancel",
+  "contact.view.delete.confirm": "Delete",
+  "contact.view.delete.description": "You can always re-add a contact later",
+  "contact.view.delete.title": "Are you sure?",
   "contacts.address": "Street address",
   "contacts.city": "City",
   "contacts.company.cityOfRegistration": "City of registration",
