+++ src/mocks/handlers.ts
@@ -2,16 +2,17 @@ import { balances } from "./handlers/home/balances";
 import { recentBills } from "./handlers/home/recent-bills";
 import { billsList } from "./handlers/bills/list";
 import { searchBills } from "./handlers/bills/search";
-import { contactList } from "./handlers/contacts/list";
 import { searchContacts } from "./handlers/contacts/search";
-import { viewContact } from "./handlers/contacts/view";
+import { getContact, getContacts } from "./handlers/contacts/get";
+import { editContact } from "./handlers/contacts/edit";
 
 export const handlers = [
   recentBills,
   balances,
   billsList,
   searchBills,
-  contactList,
   searchContacts,
-  viewContact,
+  getContact,
+  getContacts,
+  editContact,
 ];
