+++ src/mocks/handlers/contacts/search.ts
@@ -1,19 +1,19 @@
 import { http, delay, HttpResponse } from "msw";
 import { SEARCH_CONTACTS } from "@/constants/endpoints";
 import type { Contact } from "@/types/contact";
-import * as contacts from "@/mocks/handlers/contacts/list";
+import { data as contacts } from "./data";
 
 type SearchContactsResponse = {
   contacts: Contact[];
-}
+};
 
 export const data: SearchContactsResponse = {
-  contacts: contacts.data
+  contacts,
 };
 
 type SearchContactsFilter = {
   search_term?: string;
-  types?: Contact['type'][];
+  types?: Contact["type"][];
 };
 
 const filterContacts = (
@@ -22,26 +22,26 @@ const filterContacts = (
 ): SearchContactsResponse => {
   return {
     contacts: data.contacts.filter((it) => {
-      const matchesSearchTerm = !search_term ? true : Object.entries(it).some(([, value]) => {
-        if (typeof value === "object") {
-          return Object.values(it).some(
-            (innerValue) =>
-              typeof innerValue === "string" &&
-              innerValue.toLowerCase().includes(search_term.toLowerCase())
-          );
-        }
-        return (
-          typeof value === "string" &&
-          value.toLowerCase().includes(search_term.toLowerCase())
-        );
-      });
+      const matchesSearchTerm = !search_term
+        ? true
+        : Object.entries(it).some(([, value]) => {
+            if (typeof value === "object") {
+              return Object.values(it).some(
+                (innerValue) =>
+                  typeof innerValue === "string" &&
+                  innerValue.toLowerCase().includes(search_term.toLowerCase())
+              );
+            }
+            return (
+              typeof value === "string" &&
+              value.toLowerCase().includes(search_term.toLowerCase())
+            );
+          });
 
       const matchesTypes = types === undefined || types.includes(it.type);
 
-      return (
-        matchesSearchTerm && matchesTypes
-      );
-    })
+      return matchesSearchTerm && matchesTypes;
+    }),
   };
 };
 
@@ -51,7 +51,6 @@ export const searchContacts = http.post<
   SearchContactsResponse,
   typeof SEARCH_CONTACTS
 >(SEARCH_CONTACTS, async ({ request }) => {
-
   await delay(500);
 
   const { filter } = await request.json();
