+++ src/mocks/scenarios.ts
@@ -1,5 +1,5 @@
 import { emptyBillsList } from "./handlers/bills/list";
-import { emptyContactsList } from "./handlers/contacts/list";
+import { emptyContactsList } from "./handlers/contacts/get";
 
 export const scenarios = {
   empty: [emptyBillsList, emptyContactsList],
