+++ src/pages/contacts/View.tsx
@@ -2,15 +2,79 @@ import { Suspense } from "react";
 import { useNavigate, useParams } from "react-router-dom";
 import { useSuspenseQuery } from "@tanstack/react-query";
 import { FormattedMessage } from "react-intl";
-import { PencilIcon } from "lucide-react";
-
+import { PencilIcon, TrashIcon, TriangleAlert } from "lucide-react";
+import Page from "@/components/wrappers/Page";
 import Topbar from "@/components/Topbar";
 import NavigateBack from "@/components/NavigateBack";
 import PageTitle from "@/components/typography/PageTitle";
-import Page from "@/components/wrappers/Page";
+import {
+  Drawer,
+  DrawerClose,
+  DrawerContent,
+  DrawerFooter,
+  DrawerTrigger,
+} from "@/components/ui/drawer";
+import { Button } from "@/components/ui/button";
 import { getContact } from "@/services/contact";
 import Information, { Loader } from "./components/Information";
-import Delete from "./components/Delete";
+
+function Delete() {
+  return (
+    <Drawer>
+      <DrawerTrigger className="flex items-center justify-center gap-1 w-full bg-transparent">
+        <span className="text-brand-200 text-sm font-medium">
+          <FormattedMessage
+            id="contact.view.delete"
+            defaultMessage="Delete contact"
+            description="Delete contact button"
+          />
+        </span>
+        <TrashIcon className="w-3 h-3 text-brand-200" />
+      </DrawerTrigger>
+      <DrawerContent className="max-w-[375px] mx-auto bg-elevation-50">
+        <div className="flex gap-2 pt-8 px-5">
+          <TriangleAlert className="text-signal-error h-5 w-5 stroke-1" />
+
+          <div className="flex flex-col gap-1.5">
+            <span className="text-text-300 text-[18px] font-medium leading-[28px]">
+              <FormattedMessage
+                id="contact.view.delete.title"
+                defaultMessage="Are you sure?"
+                description="Delete contact title"
+              />
+            </span>
+            <span className="text-text-200 text-xs font-normal leading-[18px]">
+              <FormattedMessage
+                id="contact.view.delete.description"
+                defaultMessage="You can always re-add a contact later"
+                description="Confirm delete contact description"
+              />
+            </span>
+          </div>
+        </div>
+
+        <DrawerFooter>
+          <Button className="w-full" size="md">
+            <FormattedMessage
+              id="contact.view.delete.confirm"
+              defaultMessage="Delete"
+              description="Delete contact button"
+            />
+          </Button>
+          <DrawerClose>
+            <Button className="w-full" size="md" variant="outline">
+              <FormattedMessage
+                id="contact.view.delete.cancel"
+                defaultMessage="Cancel"
+                description="Cancel button"
+              />
+            </Button>
+          </DrawerClose>
+        </DrawerFooter>
+      </DrawerContent>
+    </Drawer>
+  );
+}
 
 function Details({ node_id }: { node_id: string }) {
   const { data } = useSuspenseQuery({
